+++ src/components/Sparkline/utilities.ts
@@ -4,7 +4,3 @@ export function getPathLength(element: SVGPathElement | null) {
   }
   return element.getTotalLength();
 }
-
-export function rgbToRgba({rgb, alpha}: {rgb: string; alpha: number}) {
-  return rgb.replace(')', `, ${alpha})`).replace('rgb', 'rgba');
-}